pyproject.toml
[project]
name = "material-agents"
version = "0.1.0"
requires-python = ">=3.11,<3.12"

dependencies = [
    "langchain-ollama",
    "langchain-core",
    "duckdb",
    "langgraph",
    "altair>=6.0.0",
    "gradio>=6.5.1",
    "vl-convert-python>=1.9.0.post1",
    "fastapi>=0.128.7",
    "uvicorn>=0.40.0",
    "pillow>=12.1.0",
    "langchain-openai>=1.1.8",
    "slowapi>=0.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "ipykernel"
]
